kernel/atomic: Update headers to use uint64_t instead of int
The underlying assembly is operating on 64 bit registers so update the
headers to be correct.

+++ src/kernel/atomic.c
@@ -2,8 +2,8 @@
 #include  <stdint.h>
 
 
-extern int _atomic_cmpxchg(uint64_t *ptr, int old, int new);
-int atomic_cmpxchg(uint64_t *ptr, int old, int new)
+extern int _atomic_cmpxchg(uint64_t *ptr, uint64_t old, uint64_t new);
+int atomic_cmpxchg(uint64_t *ptr, uint64_t old, uint64_t new)
 {
 	return _atomic_cmpxchg(ptr, old, new);
 }
@@ -12,14 +12,14 @@ int atomic_cmpxchg(uint64_t *ptr, int old, int new)
 // and dynamically pass instructions we want to be atomic
 // since the pattern is the same for these simple ops
 
-extern int _atomic_fetch_add(uint64_t *ptr, int val);
-int atomic_add(uint64_t *ptr, int val)
+extern int _atomic_fetch_add(uint64_t *ptr, uint64_t val);
+int atomic_add(uint64_t *ptr, uint64_t val)
 {
 	return _atomic_fetch_add(ptr, val);
 }
 
-extern int _atomic_fetch_or(uint64_t *ptr, int val);
-int atomic_or(uint64_t *ptr, int val)
+extern int _atomic_fetch_or(uint64_t *ptr, uint64_t val);
+int atomic_or(uint64_t *ptr, uint64_t val)
 {
 	return _atomic_fetch_or(ptr, val);
 }
